#1067 Fix Double -> BigDecimal
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@12954 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/warehouse/src/org/opentaps/warehouse/shipment/ShippingServices.java
@@ -478,9 +478,9 @@ public class ShippingServices {
         String carrierRoleTypeId = (String) context.get("carrierRoleTypeId");
         String productStoreId = (String) context.get("productStoreId");
 
-        double shippableWeight = org.ofbiz.shipment.packing.PackingServices.setSessionPackageWeights(session, packageWeights);
+        BigDecimal shippableWeight = org.ofbiz.shipment.packing.PackingServices.setSessionPackageWeights(session, packageWeights);
         session.setWeightUomId(weightUomId);
-        Double estimatedShipCost = session.getShipmentCostEstimate(shippingContactMechId, shipmentMethodTypeId, carrierPartyId, carrierRoleTypeId, productStoreId);
+        BigDecimal estimatedShipCost = session.getShipmentCostEstimate(shippingContactMechId, shipmentMethodTypeId, carrierPartyId, carrierRoleTypeId, productStoreId);
         if (UtilValidate.isNotEmpty(estimatedShipCost)) {
             session.setAdditionalShippingCharge(estimatedShipCost);
         }
